Avoid cleaning pbp file in continue mode if not measuring it
Before the pbp was always cleaned no matter if the pbp measurement was
on or off. Hence in the remote case of switching it off after having
measured it for a while, BaHaMAS was cleaning the file in any
case. This was triggering an error (trajectory not found) at a later
second resume. Now the pbp file is not cleaned if the user does not
measure it. A warninig is given, if the pbp file exists and the
measurement is turned off, though.