Improve interface to specific code w.r.t. exit codes
Since BaHaMAS enables the shell -e option and the interface was simply
calling the appropriate specific function, this call could have
terminate execution in case of non-zero exit code of the specific
function. Now the exit code is propagated up till the caller so that a
proper error handling is then possible.