Add check on number of cores in jobscript for openQCD
This is useful to have because the number of cores used at simulation
start (thermalization or new chain) could be different from the total
number of cores available per node, but this information is then to be
respected when such a simulation is continued. At the moment this is
user's responsibility and has to be done via the --coresPerNode command
line option. However, if the user forgets such an option, it is better
to catch this and remind the user than simply let the job be submitted
and then probably do not run.